move log1mexp
